Extract base dev profile output rendering
